pyproject.toml
[tool.poetry]
name = "data-structures-and-algorithms"
version = "0.1.0"
description = ""
authors = ["Mark Bell <skrambelled@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
flake8 = "^3.8.4"
pytest-watch = "^4.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
